pyproject.toml
[tool.poetry]
name = "private-gpt-backend"
version = "0.1.0"
description = ""
authors = ["vk <vineet.kumar@privado.ai>"]
readme = "README.md"
packages = [{include = "private_gpt"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pymongo = {version = "^3.12.1",extras = ["srv"]}
cryptography = "^41.0.1"
django-slack = "^5.19.0"
djongo = "^1.3.6"
python-dotenv = "^1.0.0"
mypy = "^1.3.0"
django-extensions = "^3.2.3"
openai = "^0.27.8"
django-grip = "^3.3.0"
django-cors-headers = "^4.1.0"
sseclient-py = "^1.7.2"
uwsgi = "^2.0.21"
mongodb-migrations = "^1.2.1"
presidio-anonymizer = "^2.2.341"
tiktoken = "0.5.1"
transformers = "^4.35.0"
presidio-analyzer = "^2.2.350"
spacy = "^3.7.2"

[tool.poetry.dev-dependencies]
pylint = "^2.17.4"
pre-commit = "^3.3.3"


[tool.poetry.group.dev.dependencies]
django-stubs = "^4.2.1"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
